Fix an integer signedness compiler warning.

+++ rwfile.h
@@ -78,7 +78,7 @@ public:
 	}
 	
 	/* return the current idea of the end of the file */
-	inline size_t end() const
+	inline off_t end() const
 	{
 		return write_mode ? (write_offset + filled) : write_offset;
 	}
